requirements.txt
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
llama-cpp-python
psutil

